Separate OpenGL backend in its own file
