+++ ige/src/plugin/renderer/backend.hpp
@@ -1,0 +1,17 @@
+#ifndef EA0C3563_AB58_4A44_900B_8D737ECE36A0
+#define EA0C3563_AB58_4A44_900B_8D737ECE36A0
+
+#include "ige/ecs/World.hpp"
+
+namespace ige {
+namespace renderer {
+    namespace backend {
+
+        void clear_buffers(ecs::World&);
+        void render_meshes(ecs::World&);
+
+    }
+}
+}
+
+#endif /* EA0C3563_AB58_4A44_900B_8D737ECE36A0 */
